[ALLUXIO-1910] Move AlluxioShell entry to cli package
