+++ tests/src/test/java/alluxio/shell/AbstractAlluxioShellTest.java
@@ -24,6 +24,7 @@ import alluxio.client.file.options.OpenFileOptions;
 import alluxio.exception.AlluxioException;
 import alluxio.master.LocalAlluxioCluster;
 import alluxio.security.LoginUserTestUtils;
+import alluxio.cli.AlluxioShell;
 import alluxio.util.io.BufferUtils;
 import alluxio.util.io.PathUtils;
